# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register/member": {
      "post": {
        "operationId": "AuthController_registerMember",
        "summary": "Registrasi Akun Member / Pelanggan Baru",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/register/admin-space": {
      "post": {
        "operationId": "AuthController_registerAdminSpace",
        "summary": "Registrasi Pengelola Lokasi / Admin Coworking Space",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminSpaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login Akun User (Member atau Admin Space) - Mengembalikan JWT Token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Cek Profil & Hak Akses Pengguna yang Sedang Login",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/api/diskon/check": {
      "post": {
        "operationId": "DiskonController_checkPromo",
        "summary": "Periksa Validitas & Hitung Potongan Kode Promo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckPromoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Diskon & Promo"
        ]
      }
    },
    "/api/diskon/{id}": {
      "get": {
        "operationId": "DiskonController_getById",
        "summary": "Lihat Detail Diskon Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Diskon & Promo"
        ]
      }
    },
    "/api/reservasi": {
      "post": {
        "operationId": "ReservasiController_create",
        "summary": "Buat Pemesanan Space Baru (Member)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservasiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reservasi Member"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/reservasi/my": {
      "get": {
        "operationId": "ReservasiController_getMy",
        "summary": "Lihat Status Semua Pemesanan Saya (Member)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservasi Member"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/reservasi/my/history": {
      "get": {
        "operationId": "ReservasiController_getMyHistory",
        "summary": "Lihat Histori Pemesanan Berdasarkan Bulan & Tahun (Member)",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservasi Member"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/reservasi/{id}/e-ticket": {
      "get": {
        "operationId": "ReservasiController_getETicket",
        "summary": "Cetak E-Ticket / Bukti Nota Digital Reservasi",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservasi Member"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/reservasi/{id}": {
      "get": {
        "operationId": "ReservasiController_getById",
        "summary": "Lihat Detail Reservasi Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservasi Member"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/reservasi/{id}/cancel": {
      "patch": {
        "operationId": "ReservasiController_cancel",
        "summary": "Batalkan Pemesanan (Member)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservasi Member"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/api/admin/diskon/{id}": {
      "get": {
        "operationId": "AdminController_getDiskonById",
        "summary": "Detail Data Diskon Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "AdminController_updateDiskon",
        "summary": "Update Data Kode Promo & Periode Diskon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiskonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteDiskon",
        "summary": "Hapus Kode Promo / Diskon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Space Booking API",
    "description": "Backend RESTful API untuk Sistem Reservasi Coworking Space - UKK RPL 2026/2027 SMK Telkom Malang",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterMemberDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "example": "Secret123!"
          },
          "nama_member": {
            "type": "string",
            "example": "John Doe"
          },
          "instansi": {
            "type": "string",
            "example": "Universitas Indonesia"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Sudirman No. 123, Jakarta"
          },
          "telp": {
            "type": "string",
            "example": "081234567890"
          },
          "foto": {
            "type": "string",
            "example": "member_john.jpg"
          }
        },
        "required": [
          "username",
          "password",
          "nama_member",
          "instansi",
          "alamat",
          "telp"
        ]
      },
      "RegisterAdminSpaceDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin_space1"
          },
          "password": {
            "type": "string",
            "example": "Admin123!"
          },
          "nama_coworking": {
            "type": "string",
            "example": "Moklet Hub Coworking"
          },
          "nama_pemilik": {
            "type": "string",
            "example": "Ahmad Bidin"
          },
          "telp": {
            "type": "string",
            "example": "081298765432"
          }
        },
        "required": [
          "username",
          "password",
          "nama_coworking",
          "nama_pemilik",
          "telp"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "example": "Secret123!"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "CheckPromoDto": {
        "type": "object",
        "properties": {
          "nama_diskon": {
            "type": "string",
            "example": "DISKONHEMAT20"
          }
        },
        "required": [
          "nama_diskon"
        ]
      },
      "CreateReservasiDto": {
        "type": "object",
        "properties": {
          "id_space": {
            "type": "number",
            "example": 1
          },
          "tanggal_reservasi": {
            "type": "string",
            "example": "2026-08-30"
          },
          "jam_mulai": {
            "type": "string",
            "example": "09:00"
          },
          "durasi_jam": {
            "type": "number",
            "example": 3
          },
          "id_diskon": {
            "type": "number",
            "example": 1
          },
          "kode_promo": {
            "type": "string",
            "example": "DISKONHEMAT20"
          }
        },
        "required": [
          "id_space",
          "tanggal_reservasi",
          "jam_mulai",
          "durasi_jam"
        ]
      },
...
      "UpdateDiskonDto": {
        "type": "object",
        "properties": {
          "nama_diskon": {
            "type": "string",
            "example": "PROMOAGUSTUS2026"
          },
          "persentase_diskon": {
            "type": "number",
            "example": 25
          },
          "tanggal_awal": {
            "type": "string",
            "example": "2026-08-01T00:00:00Z"
          },
          "tanggal_akhir": {
            "type": "string",
            "example": "2026-09-15T23:59:59Z"
          }
        }
      },
...
    }
  }
}